% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_event.R
\name{CheckEventCumulatively.zoo}
\alias{CheckEventCumulatively.zoo}
\title{Check the whole event window on significance.}
\usage{
CheckEventCumulatively.zoo(..., index, event.date, w.b = numeric(),
  w.a = numeric(), delta = numeric())
}
\arguments{
\item{...}{objects of class zoo, containing daily rates of return of numerous
companies.}

\item{index}{object of class zoo, containing index rates.}

\item{event.date}{date, specifying event date, which should be tested.}

\item{w.b}{numeric, parameter of time frame (window), specifying number of
days before event.}

\item{w.a}{numeric, parameter of time frame (window), specifying number of
days after event.}

\item{delta}{numeric, length of estimation period.}
}
\value{
data.frame with marked dates and test statistics
}
\description{
Check the whole event window on significance using cumulative abnormal
returns for given market index and companies.
}
\details{
Well-combined with packages tseries and yahoo finance data. Note: zoo package
should be loaded.
}
\examples{
library("tseries"); library("zoo")
date(rates)
ALV.DE <- get.hist.quote(instrument = "ALV.DE", start = "2000-01-01",
                         end = "2014-12-31", quote = "Open",
                         provider = "yahoo", retclass = "zoo")
CS.PA <- get.hist.quote(instrument = "CS.PA", start = "2000-01-01",
                         end = "2014-12-31", quote = "Open",
                         provider = "yahoo", retclass = "zoo")
CheckEventCumulatively.zoo(ALV.DE, CS.PA, index = zoo(rates[, "SXW1E"],
                           rates[, "date"]),
                           event.date = as.Date("11.09.2001","\%d.\%m.\%Y"),
                           w.b = 10, w.a = 10, delta = 110)
}

