% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_event.R
\name{CheckEventDaily.data.frame}
\alias{CheckEventDaily.data.frame}
\title{Check the event date of significance on daily basis.}
\usage{
CheckEventDaily.data.frame(rates, index.ticker = character(), event.date,
  w.b = numeric(), w.a = numeric(), delta = numeric())
}
\arguments{
\item{rates}{data.frame, which contains daily rates of return for companies
and index, as well as column with dates and name "date"}

\item{index.ticker}{character, the ticker of the index, must be a column name
of data data.frame.}

\item{event.date}{date, specifying event date, which should be tested.}

\item{w.b}{numeric, parameter of time frame (window), specifying number of
days before event.}

\item{w.a}{numeric, parameter of time frame (window), specifying number of
days after event.}

\item{delta}{numeric, length of estimation period.}
}
\value{
data.frame with marked dates and test statistics
}
\description{
Check the event date on significance on daily basis (day-by-bay) for given
market index and companies.
}
\details{
Data.frame rates should contain the column with exact name \code{date}, which
sould contain the dates, as well as column with name \code{index.ticker}
}
\examples{
data(rates)
CheckEventDaily.data.frame(rates = rates, index.ticker = "SXW1E",
                           event.date = as.Date("11.09.2001","\%d.\%m.\%Y"),
                           w.b = 10, w.a = 10, delta = 110)
}

