% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_series.R
\name{CleanSeries.zoo}
\alias{CleanSeries.zoo}
\title{Clean the series}
\usage{
CleanSeries.zoo(...)
}
\arguments{
\item{...}{objects of class zoo}
}
\value{
data.frame of series without NA's
}
\description{
Return the data.frame of series without NA's. copmlete.cases is used
}

