% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_zoo.R
\name{CombineZoosToDataframe}
\alias{CombineZoosToDataframe}
\title{Merge zoo's to the data.frame}
\usage{
CombineZoosToDataframe(...)
}
\arguments{
\item{...}{objects of class zoo}
}
\value{
data.frame, which is combined objects "zoo"
}
\description{
Return the data.frame of series of objects "zoo"
}
\examples{
library("tseries"); library("zoo")
ALV.DE <- get.hist.quote(instrument = "ALV.DE", start = "2000-01-01",
                         end = "2014-12-31", quote = "Open",
                         provider = "yahoo", retclass = "zoo")
CS.PA <- get.hist.quote(instrument = "CS.PA", start = "2000-01-01",
                         end = "2014-12-31", quote = "Open",
                         provider = "yahoo", retclass = "zoo")
CombineZoosToDataframe(ALV.DE, CS.PA)
}

