% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estudy.R
\docType{package}
\name{estudy}
\alias{estudy}
\alias{estudy-package}
\title{estudy: A package, which provides functions for event study, approach}
\description{
The package provides four catigories of basic functions: CleanSeries,
GetRates, CombineZoo and CheckEvent.
}
\section{CleanSeries functions}{

Bla-bla-bla about CleanSeries
}

\section{GetRates functions}{

Bla-bla-bla about GetRates
}

\section{CombineZoo functions}{

Bla-bla-bla Ck/fm
}

\section{CheckEvent functions}{

bla- bl
}

