% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_event.R
\name{GetNextDate}
\alias{GetNextDate}
\title{Sum of vector elements.}
\usage{
GetNextDate(date, set)
}
\description{
\code{sum} returns the sum of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the arguments
\code{...} should be unnamed, and dispatch is on the first argument.
}

