% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_event.R
\name{GetOLSAbnormals}
\alias{GetOLSAbnormals}
\title{Return the abnormals of given data.frame rates}
\usage{
GetOLSAbnormals(rates, index.ticker = character(), delta = numeric())
}
\arguments{
\item{rates}{data.frame, which contains daily rates of return for companies
and index, as well as column with dates and name "date"}

\item{index.ticker}{character, the ticker of the index, must be a column name
of data data.frame.}

\item{delta}{numeric, length of estimation period.}
}
\value{
list of abnormals, coefficients, and data.frame with initial
  (observed) returns and predicted by OLS.
}
\description{
Compute the abnormals using OLS model for rates data.frame and return the
whole parameters, such as abnormals, coefficients, and data.frame with
initial (observed) returns and predicted by OLS.
}
\details{
Data.frame rates should contain the column with exact name \code{date}, which
sould contain the dates, as well as column with name \code{index.ticker}
}
\examples{
data(rates)
GetOLSAbnormals(rates = rates, index.ticker = "SXW1E", delta = 110)
}

