% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_rates.R
\name{GetRates.numeric}
\alias{GetRates.numeric}
\title{Calculate rates of return for given series of prices.}
\usage{
\method{GetRates}{numeric}(prices)
}
\arguments{
\item{prices}{numeric vector, containing prices of the stock}
}
\value{
numeric vector of rates of return
}
\description{
Return the series of rates of return for given series of prices.
}
\details{
S3 generics, which dispatches for data.frames, zoo's and regular vector.
}
\examples{
GetRates(1:10)
}

