% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_rates.R
\name{GetRates}
\alias{GetRates}
\title{Calculate the rates of return for vector of prices.}
\usage{
GetRates(prices)
}
\arguments{
\item{prices}{numeric vector, containing prices of the stock}
}
\value{
numeric vector of the rates of return
}
\description{
Return the vector of the rates of return of given prices' vector.
}
\examples{
GetRates(1:10)
}

