% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_rates.R
\name{GetRates.zoo}
\alias{GetRates.zoo}
\title{Calculate rates of return for given series of prices.}
\usage{
\method{GetRates}{zoo}(prices)
}
\arguments{
\item{prices}{objects of class zoo, containing prices of the stock}
}
\value{
object of class zoo, which comtains rates of return
}
\description{
Return the series of rates of return for given series of prices.
}
\details{
S3 generics, which dispatches for data.frames, zoo's and regular vector.
}
\examples{
library("tseries"); library("zoo")
ALV.DE <- get.hist.quote(instrument = "ALV.DE", start = "2000-01-01",
                         end = "2014-12-31", quote = "Open",
                         provider = "yahoo", retclass = "zoo")
GetRates(ALC.DE)
}

