% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_rates.R
\name{GetRatesClosePrices.zoo}
\alias{GetRatesClosePrices.zoo}
\title{Calculate the rates of return for given series of close prices.}
\usage{
\method{GetRatesClosePrices}{zoo}(prices)
}
\arguments{
\item{prices}{objects of class zoo, containing close prices of the stock}
}
\value{
object of class zoo, which comtains the rates of return
}
\description{
Return the series of the rates of return for given series of close prices.
}
\details{
S3 generics, which dispatches for data.frames, zoo's and regular vector.
}
\examples{
library("tseries"); library("zoo")
ALV.DE <- get.hist.quote(instrument = "ALV.DE", start = "2000-01-01",
                         end = "2014-12-31", quote = "Close",
                         provider = "yahoo", retclass = "zoo")
GetRatesClosePrices(ALC.DE)
}

