% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_rates.R
\name{GetRatesOpenPrices.data.frame}
\alias{GetRatesOpenPrices.data.frame}
\title{Calculate the rates of return for given series of open prices.}
\usage{
\method{GetRatesOpenPrices}{data.frame}(prices)
}
\arguments{
\item{prices}{data.frame, in which first column is date, and the rest columns
are open prices. By default first column is dates.}
}
\value{
data.frame with respective rates of returns
}
\description{
Return the series of the rates of return for given series of open prices.
}
\details{
S3 generics, which dispatches for data.frames, zoo's and regular vector.
For data.frmae the first column will be treated as "date" column and will
be excluded from computation.
}
\examples{
library("tseries"); library("zoo")
ALV.DE <- get.hist.quote(instrument = "ALV.DE", start = "2000-01-01",
                         end = "2014-12-31", quote = "Open",
                         provider = "yahoo", retclass = "zoo")
CS.PA <- get.hist.quote(instrument = "CS.PA", start = "2000-01-01",
                         end = "2014-12-31", quote = "Open",
                         provider = "yahoo", retclass = "zoo")
prices <- CombineZoosToDataframe(ALV.DE, CS.PA)
GetRatesOpenPrices(prices)
}

