% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_series.R
\name{PlotSeries}
\alias{PlotSeries}
\title{Plot series with hatched weekends}
\usage{
PlotSeries(series, ...)
}
\arguments{
\item{series}{data.frame, with two columns, one is "date".}

\item{...}{parameters, passed to plot(...)}
}
\description{
Plot series with hatched in gray weekends
}
\details{
\code{series} should be data.frame with column "date"
}
\examples{
PlotSeries(rates[, c(1, 2)])
}

