% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estudy.R
\docType{data}
\name{rates}
\alias{rates}
\title{Dataset of 3'263 daily rates of return}
\format{A data frame with 3263 rows and 12 variables}
\usage{
data(rates)
}
\description{
Dataset contains 3'263 daily rates of return from 4.1.2000 to 12.31.2014 for
  insurance companies. Tickers of the companies represent names of the
  columns:
}
\details{
\itemize{
  \item date: dates of correspondent rates from "200-01-04" to "2014-12-31"
  \item 000810.KS: Samsung Fire & Marine Insurance Co., Ltd.
  \item AIG: American International Group, Inc.
  \item ALL: The Allstate Corporation
  \item ALV.DE: Allianz SE
  \item BRK-A: Berkshire Hathaway Inc.
  \item CS.PA: Axa SA
  \item G.MI: Assicurazioni Generali S.p.A.
  \item HNR1.HA: Hannover Rück SE
  \item MUV2.DE: Münchener Rückversicherungs-Gesellschaft Aktiengesellschaft
  \item TKOMY: Tokio Marine Holdings Inc.
  \item SXW1E: index STOXX Global 1800
}
}
\keyword{datasets}

