% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_event.R
\name{GetNextDate}
\alias{GetNextDate}
\title{Return the closest date in set, which is greater then argument date.}
\usage{
GetNextDate(date, set)
}
\arguments{
\item{date}{the date, for which the closest will be choosen}

\item{set}{the vector, which contains dates for search}
}
\value{
the closest date from \code{set}
}
\description{
Return the closest date in set, which is greater then argument date.
}
\examples{
GetNextDate(1, 1)
GetNextDate(10, 1)
}

