% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_event.R
\name{GetNextDate}
\alias{GetNextDate}
\title{The clothest date in a given set, which is greater then arument date.}
\usage{
GetNextDate(date, set)
}
\arguments{
\item{date}{the date, for which the closest will be choosen}

\item{set}{the vector, which contains dates for search}
}
\value{
the closest date from set
}
\description{
Return the closest date in set, which is greater then argument date.
}
\examples{
GetNextDate(as.Date("25.08.2005","\%d.\%m.\%Y"),
             c(as.Date("25.08.2001","\%d.\%m.\%Y"),
               as.Date("25.08.2002","\%d.\%m.\%Y"),
               as.Date("25.08.20010","\%d.\%m.\%Y")))
}

