% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_event.R
\name{GetNextDate}
\alias{GetNextDate}
\title{Return clothest to \code{date} date in the \code{set}}
\usage{
GetNextDate(date, set)
}
\arguments{
\item{date}{the single-value vector of class Date, for which the closest will
be choosen}

\item{set}{the vector of the class Date, which contains dates for search}
}
\value{
the closest date from \code{set}
}
\description{
Return the closest date in \code{set}, which is greater then argument date.
}
\details{
The \code{date} should not be greater then the largest date in \code{set}.
}
\examples{
GetNextDate(as.Date("01.05.2005","\%d.\%m.\%Y"),
    c(as.Date("01.01.2005","\%d.\%m.\%Y"), as.Date("01.08.2005","\%d.\%m.\%Y"),
    as.Date("01.03.2005","\%d.\%m.\%Y")))
}

